Update relationship methods in models for Kendaraan, MtGroup, Belanja, and Maintenance, and some code optimization

+++ app/Http/Controllers/KendaraanController.php
@@ -12,25 +12,38 @@ class KendaraanController extends Controller
 {
     /**
      * Display a listing of the resource.
+     *
+     * @return \Illuminate\Contracts\View\View
      */
     public function index()
     {
-        $kendaraans = Kendaraan::all();
-        $kendaraans = $kendaraans->map(function ($kendaraan) {
+        $kendaraans = Kendaraan::all()->map(function ($kendaraan) {
             $kendaraan->berlaku_sampai = Carbon::createFromFormat('d/m/Y', $kendaraan->berlaku_sampai)->format('Y/m/d');
             return $kendaraan;
         });
         return view('kendaraan.index', compact('kendaraans'));
     }
 
+    /**
+     * Show the form for creating a new resource.
+     *
+     * @return \Illuminate\Contracts\View\View
+     */
     public function create()
     {
         $mt_groups = MtGroup::all();
         return view('kendaraan.create', compact('mt_groups'));
     }
 
-    public function store(Request $request){
-        $request->validate([
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function store(Request $request)
+    {
+        $validatedData = $request->validate([
             'nomor_registrasi' => 'required|string|max:255|unique:tbl_kendaraan',
             'merk_kendaraan' => 'required|string|max:255',
             'jenis_kendaraan' => 'required|string|max:255',
@@ -40,50 +53,66 @@ class KendaraanController extends Controller
             'berlaku_sampai' => 'required|date_format:d/m/Y',
             'mt_group' => 'required|integer',
         ],[
-            'required' => 'Kolom :attribut wajib diisi.',
-            'integer' => 'Kolom :attribut harus berupa angka.',
-            'date_format' => 'Kolom :attribut tidak sesuai format d/m/Y.',
+            'required' => 'Kolom :attribute wajib diisi.',
+            'integer' => 'Kolom :attribute harus berupa angka.',
+            'date_format' => 'Kolom :attribute tidak sesuai format d/m/Y.',
             'unique' => 'Nomor registrasi sudah digunakan.',
         ]);
 
-        $data_kendaraan = $request->except('mt_group');
-        
-        
-        $data_kendaraan['berlaku_sampai'] = Carbon::createFromFormat('d/m/Y', $request->berlaku_sampai)->format('d/m/Y');
-        $kendaraan = Kendaraan::create($data_kendaraan);
-        
-        $data_maintenance['nomor_registrasi'] = $request->nomor_registrasi;
-        $data_maintenance['mt_group'] = $request->mt_group;
-        $data_maintenance['_token'] = $request->_token;
-        $data_maintenance = Maintenance::create($data_maintenance);
+        $validatedData['berlaku_sampai'] = Carbon::createFromFormat('d/m/Y', $validatedData['berlaku_sampai'])->format('d/m/Y');
+        $kendaraan = Kendaraan::create($validatedData);
+
+        $maintenanceData = [
+            'nomor_registrasi' => $request->nomor_registrasi,
+            'mt_group' => $request->mt_group,
+        ];
+        Maintenance::create($maintenanceData);
 
         if ($kendaraan->wasRecentlyCreated) {
             return redirect()->route('kendaraan.index')->with('success', 'Data berhasil disimpan.');
         } else {
             return redirect()->route('kendaraan.index')->with('error', 'Terjadi kesalahan saat menyimpan data.');
         }
-        return redirect()->back();
     }
 
-    public function printAll(){
+    /**
+     * Display the specified resource.
+     *
+     * @return \Illuminate\Contracts\View\View
+     */
+    public function printAll()
+    {
         return view('kendaraan.printAll');
     }
 
-    public function edit($id){
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Contracts\View\View
+     */
+    public function edit($id)
+    {
         $kendaraan = Kendaraan::find($id);
-        $maintenance = Maintenance::join('tbl_mt_group', 'tbl_maintenance.mt_group', '=', 'tbl_mt_group.id')
-        ->where('nomor_registrasi', $kendaraan->nomor_registrasi)
-        ->select('tbl_maintenance.nomor_registrasi','tbl_mt_group.id', 'tbl_mt_group.nama_group')
-        ->first();
+        $maintenance = Maintenance::where('nomor_registrasi', $kendaraan->nomor_registrasi)
+            ->join('tbl_mt_group', 'tbl_maintenance.mt_group', '=', 'tbl_mt_group.id')
+            ->select('tbl_maintenance.nomor_registrasi', 'tbl_mt_group.id', 'tbl_mt_group.nama_group')
+            ->first();
     
         $mt_groups = MtGroup::all();
-
-        // dump($maintenance);
     
         return view('kendaraan.edit', compact('kendaraan', 'mt_groups', 'maintenance'));
     }
 
-    public function update(Request $request, $id){
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function update(Request $request, $id)
+    {
         $validatedData = $request->validate([
             'nomor_registrasi' => 'required|string|max:255',
             'merk_kendaraan' => 'required|string|max:255',
@@ -98,26 +127,29 @@ class KendaraanController extends Controller
             'date_format' => 'Kolom :attribute tidak sesuai format d/m/Y.',
         ]);
 
-
         $validatedData['berlaku_sampai'] = Carbon::createFromFormat('d/m/Y', $validatedData['berlaku_sampai'])->format('d/m/Y');
         
         $kendaraan = Kendaraan::find($id);
-        $kendaraan->update($request->all());
+        $kendaraan->update($validatedData);
 
         return redirect()->route('kendaraan.index')->with('success', 'Data berhasil diperbarui.');
-
-        // return redirect()->back();
     }
 
-    public function destroy($id){
-
-        // TODO: Delete data kendaraan
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\RedirectResponse
+     */
+    public function destroy($id)
+    {
         $kendaraan = Kendaraan::find($id);
 
-        $kendaraan->delete();
-
-        return redirect()->route('kendaraan.index')->with('success', 'Data berhasil dihapus.');
-
-        return redirect()->back();
+        if ($kendaraan) {
+            $kendaraan->delete();
+            return redirect()->route('kendaraan.index')->with('success', 'Data berhasil dihapus.');
+        } else {
+            return redirect()->route('kendaraan.index')->with('error', 'Data tidak ditemukan.');
+        }
     }
 }
